Peta el parse de la fecha
