+++ sesion08Entregable/app/src/main/res/values-w820dp/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="jugar">Jugar</string>
+</resources>
