+++ sesion08Entregable/app/src/main/res/values/strings.xml
@@ -10,4 +10,6 @@
     <string name="guardadas">Últimas partidas jugadas</string>
     <string name="sinpartidasguardadas">No se encontraron partidas anteriores</string>
     <string name="partidasguardadas">Last Saved Matches</string>
+    <string name="jugar">Jugar</string>
+    <string name="registrarse">Registrar Nuevo Jugador</string>
 </resources>
